+++ lizard_portal/media/lizard_portal/lizard/popup/FeatureInfo.js
@@ -10,87 +10,35 @@
     width: 300,
     init_background: null,
     statics: {
-      show: function(record) {
+      show: function(records, workspaceitem) {
+        var data, html, record, tpl;
+        record = records[0];
+        data = [];
+        Ext.Object.each(record.data, function(key, value) {
+          return data.push({
+            key: key,
+            value: value
+          });
+        });
+        tpl = new Ext.XTemplate('<div class="lizard">', '<h2>Kaartlaag: {layer_name}</h2>', '<table>', '<tpl for="fields">', '<tr><td>{key}</td><td>{value}</td></tr>', '</tpl></table></div>');
+        html = tpl.applyTemplate({
+          layer_name: workspaceitem.get('title'),
+          fields: data
+        });
         return Ext.create('Ext.window.Window', {
-          title: 'Feature Info',
-          is_feature_info: true,
-          items: Ext.create('Lizard.popup.FeatureInfo', {
-            record: record
-          })
+          title: 'Info',
+          popup_type: 'feature_info',
+          items: [
+            {
+              xtype: 'panel',
+              width: 400,
+              html: html
+            }
+          ]
         }).show();
       }
     },
-    items: [
-      {
-        fieldLabel: 'Achtergrondkaart',
-        name: 'base_layer',
-        displayField: 'title',
-        valueField: 'plid',
-        xtype: 'combo',
-        queryMode: 'local',
-        autoSelect: true,
-        typeAhead: false,
-        value: this.startValue,
-        minChars: 0,
-        forceSelection: true,
-        allowBlank: false,
-        width: 200,
-        store: {
-          pageSize: 10000,
-          model: 'Lizard.model.WorkspaceItemModel',
-          proxy: {
-            type: 'ajax',
-            url: '/workspace/api/layer_view/?_accept=application%2Fjson',
-            extraParams: {
-              filter: '[{"property": "is_base_layer", "value": true}]'
-            },
-            reader: {
-              type: 'json',
-              root: 'data'
-            }
-          }
-        }
-      }
-    ],
-    bbar: [
-      '->', {
-        text: 'Annuleren',
-        handler: function(btn, event) {
-          var window;
-          window = this.up('window');
-          return window.close();
-        }
-      }, {
-        text: 'OK',
-        handler: function(btn, event) {
-          var base_layer, form, index, values, window;
-          form = this.up('form').getForm();
-          if (form.isValid()) {
-            values = form.getValues();
-            index = form.findField('base_layer').store.find('plid', values.base_layer);
-            base_layer = form.findField('base_layer').store.getAt(index);
-            Lizard.CM.setContext({
-              background_layer: base_layer.raw
-            });
-            window = this.up('window');
-            return window.close();
-          } else {
-            return Ext.MessageBox.alert('Invoer fout', 'Kies geldige periode');
-          }
-        }
-      }
-    ],
-    afterRender: function() {
-      var form;
-      form = this.getForm();
-      form.findField('base_layer').store.load();
-      if (this.init_background_id === null) {
-        this.init_background_id = Lizard.CM.getContext().background_layer.id;
-      }
-      if (this.init_background_id) {
-        return form.findField('base_layer').setValue(this.init_background_id);
-      }
-    }
+    items: [{}]
   });
 
 }).call(this);